third_party/blink/web_tests/virtual/popover-anchor-relationships-disabled/external/wpt/html/semantics/popovers/imperative-invokers-expected.txt

This is a testharness.js-based test.
[FAIL] showPopover(popover)
  assert_true: popovers should be related to each other expected true got false
[FAIL] showPopover(contained)
  assert_true: popovers should be related to each other expected true got false
[FAIL] togglePopover(popover)
  assert_true: popovers should be related to each other expected true got false
[FAIL] togglePopover({force, popover})
  assert_true: popovers should be related to each other expected true got false
Harness: the test ran to completion.
